2018-12-08 修改Cache增加PersistentAllocator的支持

+++ utilities/nvm_write_cache/nv_range_tab.h
@@ -42,7 +42,7 @@ public:
 
     bool equals(NvRangeTab &b);
     char* raw_;
-    p<uint64_t> offset_;
+    p<int> offset_;
 
     p<uint64_t> hash_;
